ZenoDcimManager.Tests/UserContext/Handlers/UserHandlerTests.cs
using ZenoDcimManager.Domain.UserContext.Commands;
using ZenoDcimManager.Domain.UserContext.Handlers;
using ZenoDcimManager.Domain.UserContext.Repositories;
using ZenoDcimManager.Domain.UserContext.Services;
using ZenoDcimManager.Shared.Services;
using ZenoDcimManager.Tests.UserContext.Repositories;
using ZenoDcimManager.Tests.UserContext.Services;
using Microsoft.VisualStudio.TestTools.UnitTesting;

namespace ZenoDcimManager.Tests.UserContext.Handlers
{
    [TestClass]
    public class UserHandlerTests
    {
        private readonly IUserRepository _userRepository;
        private readonly IEmailService _emailService;
        private readonly ICryptoService _cryptoService;
        private UserHandler _handler;
        public UserHandlerTests()
        {
            _userRepository = new FakeUserRepository();
            _emailService = new FakeEmailService();
            _cryptoService = new FakeCryptoService();
            _handler = new UserHandler(_userRepository, _emailService, _cryptoService);
        }
        [TestMethod]
        [TestCategory("Handlers")]
        public void ShouldCreateUserWhenCommandIsValid()
        {
            var command = new CreateUserCommand("Marcelo", "Borba", "marcelo@marcelo.com", "123456", "123456", 1, true);
            _handler.Handle(command);
            Assert.AreEqual(_handler.Valid, true);
        }

        [TestMethod]
        [TestCategory("Handlers")]
        public void ShouldCreateUserWhenCommandIsInvalid()
        {
            var command = new CreateUserCommand("Marcelo", "Borba", "marcelo@marcelo.com", "123456", "123456", 8, true);
            _handler.Handle(command);
            Assert.AreEqual(_handler.Invalid, true);
        }

        [TestMethod]
        [TestCategory("Handlers")]
        public void ShouldAddANewUserOnRepository()
        {
            // var userCount = _userRepository.List().Count;
            // var command = new CreateUserCommand("Marcelo", "Borba", "marcelo@marcelo.com", "123456", "123456", 1, true);
            // var handler = new UserHandler(_userRepository, _emailService);
            // handler.Handle(command);
            // var userCountUpdated = _userRepository.List().Count;
            // Assert.AreNotEqual(userCount, userCountUpdated);
            Assert.IsTrue(true);
        }
    }
}
